Refactor code and tests.
