+++ metrics_utility/library/__init__.py
@@ -1,6 +1,7 @@
 from . import collectors, dataframes, extractors, instants, package, reports, storage
 from .csv_file_splitter import CsvFileSplitter
-from .utils import last_gather, lock, save_last_gather, tempdir
+from .lock import lock
+from .utils import last_gather, save_last_gather, tempdir
 
 
 __all__ = [
